requirements.txt
python-telegram-bot>=22.0,<23
python-dotenv>=1.0
pytz>=2024.1
